Keep offset and readahead size data type uniform
Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

+++ db/version_set.cc
@@ -1198,7 +1198,7 @@ void LevelIterator::SetFileIterator(InternalIterator* iter) {
 
   if (read_options_.reuse_internal_auto_readahead_size && old_iter != nullptr &&
       file_iter_.iter() != nullptr) {
-    uint64_t prev_readahead_size = old_iter->GetInternalCurrentReadAheadSize();
+    size_t prev_readahead_size = old_iter->GetInternalCurrentReadAheadSize();
     // file_iter_ points to BlockBasedTableIterator for L1 - Ln level.
     file_iter_.SetInternalInitialReadAheadSize(prev_readahead_size);
   }
